apply suggestions and move tasks test
